+++ agents_infra/agents_infra/agents/base.py
@@ -22,9 +22,9 @@ import urllib2
 from dateutil import parser
 from urlparse import urljoin
 
+from ..managers.health_server_manager import HealthServerManager
 from ..utils import LOG_CONFIG_PATH, maybe_log_message
 from ..utils.configtools import get_config_option, parse_csv_list
-from ..utils.health_server_manager import HealthServerManager
 
 PROTOCOLS = ('tcp', 'udp')
